Fallback with these mixes in turn.
The signature of the `opp-s-wrap-maybe' is

  opp-s-wrap-maybe t1 t2 s1-0 s2-0 s1-1 s2-1 ...

If s1-0 and s2-0 both found in the BUFFER, replace them with t1 and
t2. If not found, try s1-1 and s2-1, ... in turn.
For example, `def-opp-surround-pair { } B` expands into

  def-opp-surround "{" "} " " }" "{" "}" ...

Eventually `opp-s-wrap-maybe' will be called something like this.

  opp-s-wrap-maybe "<" ">" "{ " " }" "{" "}"

In this case, (assume the cursor is inside of the BUFFER)
if the BUFFER is
a) "buffer"     then the ruselt will be "buffer". (does nothing)
b) "{ buffer }" then the result will be "<buffer>".
c) "{buffer}"   then the result will be "<buffer>", too.

I want to implement and experiment this naive behavior.

Signed-off-by: Takeshi Banse <takebi@laafc.net>